[UPDATE] Conexão com o banco de dados, implementação de testes e ajuste no método alterar

+++ backend/Uniftec.ProjetoWeb.SocialTec/Uniftec.ProjetoWeb.SocialTec.Repository/Repository/PublicacaoRepository.cs
@@ -26,24 +26,28 @@ namespace Uniftec.ProjetoWeb.SocialTec.Repository.Repository
                     cmd.Parameters.AddWithValue("id", publicacao.Id);
                     cmd.Parameters.AddWithValue("usuario", publicacao.Usuario);
                     cmd.Parameters.AddWithValue("descricao", publicacao.Descricao);
-                    cmd.Parameters.AddWithValue("datapublicacao", publicacao.DataPublicacao);
+                    cmd.Parameters.AddWithValue("datapublicacao", DateTime.Now);
                     Console.WriteLine(cmd.ExecuteNonQuery());
 
-                    cmd.Parameters.Clear();
-                    cmd.CommandText = "Delete from publicacaomidia where idpublicacao=@id";
-                    cmd.Parameters.AddWithValue("id", publicacao.Id);
-                    cmd.ExecuteNonQuery();
-
-                    foreach (var midia in publicacao.UrlsMidia)
+                    if (publicacao.UrlsMidia.Count != 0)
                     {
                         cmd.Parameters.Clear();
-                        cmd.CommandText = @"INSERT INTO public.publicacaomidia 
+                        cmd.CommandText = "Delete from publicacaomidia where idpublicacao=@id";
+                        cmd.Parameters.AddWithValue("id", publicacao.Id);
+                        cmd.ExecuteNonQuery();
+
+                        foreach (var midia in publicacao.UrlsMidia)
+                        {
+                            cmd.Parameters.Clear();
+                            cmd.CommandText = @"INSERT INTO public.publicacaomidia 
                                                 (idpublicacao, id, url)
                                           VALUES(@idpublicacao, @id, @url)";
-                        cmd.Parameters.AddWithValue("idpublicacao", publicacao.Id);
-                        cmd.Parameters.AddWithValue("id", Guid.NewGuid());
-                        cmd.Parameters.AddWithValue("url", midia);
-                        cmd.ExecuteNonQuery();
+                            cmd.Parameters.AddWithValue("idpublicacao", publicacao.Id);
+                            cmd.Parameters.AddWithValue("id", Guid.NewGuid());
+                            cmd.Parameters.AddWithValue("url", midia);
+                            cmd.ExecuteNonQuery();
+                        }
+
                     }
                 }
             }
